requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.44.0
django-countries==7.2.1
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-heroku==0.3.1
django-storages==1.11.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djongo==1.3.4
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
install==1.3.4
mysqlclient==2.0.3
numpy==1.20.3
oauthlib==3.1.0
Pillow==8.1.2
psycopg2-binary
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.3
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.2.4
stripe==2.57.0
urllib3==1.26.4
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
